+++ CHANGELOG.md
@@ -5,10 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.1.0] - 2020-05-09
 
 ### Added
 - Added `line-[style]` and `line-[color]` utilities to go with Tailwind’s text decoration utilities (`underline` and `line-through`); they can be customized with the `textDecorationStyle` and `textDecorationColor` theme keys
+
+### Fixed
+- Fixed an issue when using a font size that includes a default line height in a text style (e.g. `['16px', '24px']`), which is supported since Tailwind 1.3
 
 ## [3.0.1] - 2020-02-13
 
@@ -96,7 +99,8 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 Initial release
 
-[Unreleased]: https://github.com/benface/tailwindcss-typography/compare/v3.0.1...HEAD
+[Unreleased]: https://github.com/benface/tailwindcss-typography/compare/v3.1.0...HEAD
+[3.1.0]: https://github.com/benface/tailwindcss-typography/compare/v3.0.1...v3.1.0
 [3.0.1]: https://github.com/benface/tailwindcss-typography/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/benface/tailwindcss-typography/compare/v2.2.0...v3.0.0
 [2.2.0]: https://github.com/benface/tailwindcss-typography/compare/v2.1.1...v2.2.0